Refactored code
changed the modal of the tovit to be used as modal container for every content when needed.
change the user last post time with he submit new post.
moved some styles to new file to make the style more readyable
changed the db to be able to handle multi querys at once